refactor: divide select page components
